requirements.txt
accelerate==0.30.1
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
albucore==0.0.15
albumentations==1.4.15
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
datasets==2.20.0
diffusers==0.28.0
dill==0.3.8
easydict==1.13
einops==0.8.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
fastapi==0.112.4
ffmpy==0.4.0
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.2.0
google-ai-generativelanguage==0.6.9
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.1
googleapis-common-protos==1.65.0
gradio==4.31.5
gradio_client==0.16.4
groundingdino-py @ git+https://github.com/giswqs/GroundingDINO@2e4e563c96f30d97425cd9f0fb6f740d3f66349a
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.24.7
humanfriendly==10.0
idna==3.8
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
insightface==0.7.3
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.8.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
onnx==1.16.2
onnxruntime==1.17.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.7
packaging==24.1
pandas==2.2.2
peft==0.11.1
pillow==10.2.0
platformdirs==4.3.3
prettytable==3.11.0
proto-plus==1.24.0
protobuf==5.28.1
psutil==6.0.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pydantic==2.9.1
pydantic_core==2.23.3
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
ruff==0.6.5
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.0
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.38.5
supervision==0.6.0
sympy==1.12
threadpoolctl==3.5.0
tifffile==2024.8.30
timm==1.0.9
tokenizers==0.15.2
tomli==2.0.1
tomlkit==0.12.0
torch==2.4.1+cu121
torchvision==0.19.1+cu121
tqdm==4.66.5
transformers==4.37.2
triton==3.0.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.30.6
wcwidth==0.2.13
websockets==11.0.3
xxhash==3.5.0
yapf==0.40.2
yarl==1.11.1
zipp==3.20.2
